Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> If you update x/crypto to at least https://go.googlesource.com/crypto/+/ea8f1a30c4438cc8b13f05538385ad8dc6049b43, there should be no need to do anything else than adding a case in the PKCS#8 switch, because the x/crypto/ed25519 types get aliased to the new crypto/ed25519 ones.
> 
> Then you can keep the ASN.1 manipulations as fallbacks for pre-Go 1.13, and eventually just remove them.

We're actually just ad-hoc vendoring x/crypto right now. :-(

Originally it was due to dependencies being a headache (GOPATH doesn't work great for non-Go projects that happen to have some Go in them) and we needed to patch x/crypto/ed25519 to add NewKeyFromSeed. Now that's all upstreamed and Go modules are a thing. I just need to check with Chrome infra about how best to deal with the module cache and still haven't gotten around to it.

But our documented policy is that we only try to support the latest stable Go 1.13, so I'm not too worried about having to keep both around.

Patch-set: 2