Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Hey !
> > 
> > As discussed in a private tweet with Adam a couple of months ago, here is a simple fix for removing some devs-privacy-hurting strings in release builds of BoringSSL. 
> > 
> > Let me know what you think about it.
> > 
> > Best, 
> > 
> > Pascal
> 
> We use the filenames to diagnose errors all the time in logs from release builds. The full path isn't necessary, but I don't think we can remove this without some other way to identify the file.

On my builds with GCC and Clang, it reports the relative path passed in on the command-line anyway. Not sure about MSVC, but it seems their __FILE__ macro is controllable, so you may be able to solve it at the build level. (Doing it at the build level is also more robust as lots of things sample __FILE__ and __LINE__.)

https://docs.microsoft.com/en-us/cpp/build/reference/fc-full-path-of-source-code-file-in-diagnostics?view=vs-2019
https://docs.microsoft.com/en-us/cpp/preprocessor/predefined-macros?view=vs-2019

Patch-set: 1
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>