Update patch set 2
Patch Set 2:

> where would you recommend potentially placing this, in that case (since libcrypto can't depend on libdecrepit)?

So, I think libdecrepit is the right place for CFB mode. As Adam notes, this is a very obscure cipher. We have many consumers of BoringSSL and those consumers should not have to pay costs (both in binary size as well as the security analysis cost of more code being potentially reachable) of such modes. That means we don't want it in libcrypto. This also means EVP_get_cipherbyname can't depend on it. Callers that use EVP_get_cipherbyname have to pull in *all* ciphers attached to that function because the static linker can't tell it's unused.

This came up in https://boringssl-review.googlesource.com/c/boringssl/+/33984 as well, where we also declined patches which linked libcrypto to libdecrepit. Really the problem is that EVP_get_cipherbyname is a terrible API, but we're stuck with that one. :-/

(Admittedly, libcrypto includes other undesirable things like EVP_rc2_cbc. That one is, sadly, necessary to support PKCS#12 files and so surprisingly not obscure. Though perhaps we should have declined to add it to EVP_get_cipherbyname in that other CL for similar reasons, just to limit it's scope...)


I don't particularly mind the implementation of EVP_aes_256_cfb128 itself since it shares code with EVP_aes_128_cfb128, but please add tests.

Patch-set: 2