Update patch set 1
Abandoned

We have highlighted the introduction to BoringSSL from the home page several times to contacts at IBM: “BoringSSL is a fork of OpenSSL that is designed to meet Google's needs. Although BoringSSL is an open source project, it is not intended for general use, as OpenSSL is.‭”

I'm afraid that taking on additional platforms is a significant commitment, even more so when it's a different endianness from the rest of the world. There has to be a business case for Google to do this and none has been made for the s390x platform.

Patch-set: 1
Status: abandoned
Tag: autogenerated:gerrit:abandon