Update patch set 1
Patch Set 1:

After some further discussion, it seems like it'd be better to make this output JSON rather than having the ad hoc format I created here. I wrote a change for this. Would it be the proper workflow to push the change to this review or should I make a new commit with only that change?

Patch-set: 1
CC: Gerrit User 9735 <9735@66b711a4-aa9b-3dac-9062-916540830212>