Update patch set 1
Patch Set 1:

The reason for not using static inline is that it's, strictly speaking, an ODR violation (undefined behavior) in C++. Additionally, should the compiler need to emit an out-of-line version, static inline forces the compiler to emit multiple copies while plain inline deduplicates in C++.

In particular, these symbols are public API, so they may be used in other inline C++ functions in calling code, so we don't know if they, e.g., take a function pointer or something else odd. (Indeed passing sk_FOO_free around as a function pointer is not all that unlikely.)

This was mostly a theoretical issue (UB issues around ODR are mostly a spec translation of the compiler/linker split), but tempting the compiler with UB is generally not the best plan.

I'd like to understand better what's going on with this LLVM warning. The C files shouldn't be creating symbols at all, and there's nothing wrong with one translation unit defining a file-local FOO and another defining an exported FOO. The file-local one is file-local for a reason.

Patch-set: 1