Update patch set 13
Patch Set 13:

I think this is ready to go but I've asked davidben to take a look over because changes to the PRNG are esp sensitive. In the follow-on change I've gotten the Android + FIPS build working in our CMake system so that we can setup a builder for this configuration.

Patch-set: 13
Reviewer: Gerrit User 5005 <5005@66b711a4-aa9b-3dac-9062-916540830212>