Update patch set 4
Patch Set 4:

> (One last thought: if there is no standard, maybe the context string could be less BoringSSL-specific so that others might adopt this? Not sure if it's worth it, but something to ponder for a second.)

That's a good point. There're two sources of BoringSSL-specific strings, the label itself and the name of the curve. Making the label less BoringSSL-specific SGTM. Do you think we should ditch the curve name or just assume "P-256" and "P-384" are reasonably okay?

TBH the label was mostly because it was free. It seems odd for a protocol to manage to misuse the secret, but *only* by deriving a key for each curve.

Patch-set: 4
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>