urandom: drop extra getrandom non-blocking call.
When probing for whether the kernel supports |getrandom| we learn
whether or not the pool is initialised. If not, then |wait_for_entropy|
does another probe before printing a warning and doing a blocking call.

This change removes that second probe because it'll always be called
nearly immediately after the first one and the chance that the kernel
managed to initialise the pool in that time is trivial.

Considering cases to justify the claim that it'll be called immediately
after the first probe:

FIPS:
  Have RDRAND: only a non-blocking draws are ever done, so moot
  No RDRAND:   the first draw is blocking, so true.
Non-FIPS:
  Have RDRAND: the first draw is blocking, so true.
  No RDRAND:   the first draw is blocking, so true.

Change-Id: I55c56c3163c1b6cbba11afe1afa43d5126a7c863