Update patch set 4
Patch Set 4:

(4 comments)

> Patch Set 4:
> 
> I see that SSL_get_shared_sigalgs used to be present and was removed several
> years ago in https://boringssl.googlesource.com/boringssl/+/ba16a1e405c617f4179bd780ad15522fb25b0a65%5E!
> so I wonder if you all would consider perhaps adding back the actual
> implementation to enable actual compatibility in Electron with the newly
> added methods in Node.js?

Nah, that was removed for a reason. It would require use to retain extra per-connection memory past the end of the handshake, as well as calculate an intermediate state that we don't use to begin with. On a server with many connections, per-connection memory adds up.

Patch-set: 4
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>