Add standalone CMake build to generate_build_files.py
This change adds a 'cmake' build type to generate_build_files.py, which
creates a standalone, CMake build without (most) of the complexities of
the full development CMake build.

(This is just a sketch for now. Some deduplication of the CMake is
called for before landing, presuming that this is useful.)

Change-Id: I1f91a8905cabaedf8f9ec588da5784edd67bfed0