Update patch set 3
Patch Set 3:

Can we enumerate the possible side-effects of certificate verification for clients that set SSL_VERIFY_NONE?  I can think of two:

1. The result is observable through SSL_get_verify_result().
2. Callbacks installed with SSL_/SSL_CTX_set_cert_verify_callback(), SSL_/SSL_CTX_set_verify() will be called.

Suppose that we try to disallow calls to these functions if SSL_VERIFY_NONE.  (The signatures of these functions make this tricky to do: calling abort() would work, though it's pretty heavy-handed.)

But if we can get away with it, then I would think the additional step of not performing certificate verification in the SSL_VERIFY_NONE case would be provably a no-op.

This, I take it, is the end-state that you want?  I'm not sure we can get there, or that it's worth the effort, but I can certainly give it a whirl.  (I am pretty sure this will require invasive changes to third-party software, though arguably worthwhile ones.)

That said:

Even if we did all this, I would *still* want to land the present change, because SSL_VERIFY_NONE is the most straightforward way to spell "this code doesn't authenticate", and in particular it is a lot more intuitive to an auditor than your proposed alternative, which I think is this?

  SSL_CTX_set_verify(ctx, SSL_VERIFY_PEER, nullptr);
  SSL_CTX_set_cert_verify_callback(ctx, []() { return 1; }, nullptr);

I'm none too wedded to any of this, but that's my thinking.

Patch-set: 3
Reviewer: Gerrit User 5645 <5645@66b711a4-aa9b-3dac-9062-916540830212>