Update patch set 2
Patch Set 2: Code-Review+2

Hrmf. The aarch64 ABI is perfectly capable of supporting alloca, so the compiler really should be able to figure this out. :-/

This works, but would it work to just change it to alignas(16)? This is testing when all the inputs are *not* aligned and even being misaligned mod 2 implies you are misaligned mod 64.

Patch-set: 2
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=+2