Proxy: send whole SSL records through the handshaker.
In split handshake tests, it is already the case that the handshaker
must signal the proxy when it wants to read more data.  But there was
not a lot of specificity about exactly how much data would be read.

The addition of early data to TLS 1.3 requires this to be nailed down,
in order that the handshaker should receive only the handshake, and
not the early data.

This commit changes the handshaker to read exactly one SSL record and
then stop, when it is asked to read.  The pattern of I/O operations
remains undefined.

Change-Id: I30f58e57fc5ebff3f7c7ef8482cc629e42fef6a4