Calling RtlGenRandom on Desktop Apps and BCryptGenRandom on UWP Apps
This CL conditions the API used for generating entropy on the targeted Windows partition.
BCryptGenRandom will be used on UWP apps and RtlGenRandom will be used on non-UWP apps (Desktop).
On non-UWP configurations, RtlGenRandom is used instead of BCryptGenRandom to avoid accessing resources that may be unavailable inside the Chromium sandbox.

Bug: 307
Change-Id: I49f445198b7b4f300a752f45e221a2875d17099e