Update patch set 1
Patch Set 1:

(2 comments)

We first have the frontend read the ClientHello, and then we cut to the handshaker, then we cut back after ServerHello..Finished, right? Since the frontend is parsing the ClientHello, it already knows where the boundary between ClientHello and early data is.

Is this only needed for the 0-RTT + HelloRetryRequest (implies 0-RTT reject) case? There I could see this being a nuisance, because we cut before resolving HRR rather than after. (After would be attractive in a number of ways, but requires more of the TLS config live on the frontend.)

Sending one record at a time works for tests, but it is a little odd when it comes to actually implementing this since we wouldn't want to pay a roundtrip for every record. The more general streaming solution would be for the frontend to stream data, get a read ACK (or, if unbounded, several), and be told "hey, I only consumed N bytes, please keep processing the rest". That wouldn't even require parsing records.

For specifically the HRR case, the frontend can also tell based on record type, though it is similarly making assumptions about the protocol.

Patch-set: 1
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>