Support building as a Windows UWP app.
RtlGenRandom is a legacy API that might be altered and is unavailable
for UWP apps. BCryptGenRandom is the recommended API for generating
random numbers on UWP. Unfortunately, BCryptGenRandom doesn't have the
same performance level as RtlGenRandom. In order to allow UWP apps to
benefit from BoringSSL, this change replaces RtlGenRandom with
BCryptGenRandom only for UWP builds. To the best of my knowledge,
BCryptGenRandom with BCRYPT_USE_SYSTEM_PREFERRED_RNG should behave
similarly to RtlGenRandom.

Bug: 307
Change-Id: I49f445198b7b4f300a752f45e221a2875d17099e