Update patch set 2
Patch Set 2:

(2 comments)

That's right.  I've updated the commit message.

You're right that paying a round-trip for each record is not ideal, but in tests, we are already doing that.  Production code could do something different.  Do you think it's important to make the tests work in the same way?

Patch-set: 2