Update patch set 1
Patch Set 1: Commit-Queue+1

(CLA confirmed.)

The level of arcania to get entropy on Windows is a little crazy :(

Obviously we don't have any builders nor consumers who exercise this because this is the first time that I'm aware that anyone has hit this. Therefore there's nothing to stop us regressing support for UWP in the future. Typically we are hesitant to support things that we can't support property but, I think this counts as "Windows", and we're supposed to work on Windows.

I'll leave this a little while because David or Ryan might have good ideas. Mostly I'm assuming from your email address that this is correct because I know nothing!

Patch-set: 1
Reviewer: Gerrit User 5005 <5005@66b711a4-aa9b-3dac-9062-916540830212>
Label: Commit-Queue=+1