Update patch set 2
Patch Set 2:

> Patch Set 1: Commit-Queue+1
> 
> (CLA confirmed.)
> 
> The level of arcania to get entropy on Windows is a little crazy :(
> 
> Obviously we don't have any builders nor consumers who exercise this because this is the first time that I'm aware that anyone has hit this. Therefore there's nothing to stop us regressing support for UWP in the future. Typically we are hesitant to support things that we can't support property but, I think this counts as "Windows", and we're supposed to work on Windows.
> 
> I'll leave this a little while because David or Ryan might have good ideas. Mostly I'm assuming from your email address that this is correct because I know nothing!

The context for this is https://crbug.com/boringssl/307. This CL is gated on UWP, so it won't directly run into Chrome's sandboxing needs, but I would like a clear answer as to what BCryptGenRandom with that flag actually does. If nothing else, that way we can document why we still have the RtlGenRandom codepath, rather than just sending everything through BCryptGenRandom.

Patch-set: 2